before/app/requirements.txt
pytest==8.3.5
pytest-cov==6.1.1
celery==5.5.1
celery[sqlalchemy]==5.5.1
redis==5.2.1
fastapi==0.115.6
uvicorn==0.34.0
flower==2.0.1
pydantic==2.10.5
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pandas==2.2.3
pymupdf==1.25.5
numpy==2.2.4
matplotlib==3.10.1
httpx==0.28.1
boto3==1.37.26
colorlog==6.9.0
chardet==5.2.0
xlrd==2.0.1
openpyxl==3.1.5